collector/requirements.txt
aiogram==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0